Small correction to the definition of fmap for FreeRigidSelective
The given `f: A => B` was being applied to the left hand side of the inner
`Either[w.Source, A]` in the `Select` case, resulting in a runtime
`ClassCastException` when attempting to construct a program utilising `map`: the
`Unit` value at the leaf of the structure was used as the argument to `f`, when
a value of type `A` is expected. The reason this isn't caught by the compiler is
that the pattern-matched type was unchecked.  This unchecked refinement type is
still unchecked.

This commit:

 - corrects the `Source` type member to align with the pattern-matched `Select`

 - applies the `f: A => B` to the "result" of the inner `Select` by mapping over
   the right-hand Either value

 - swaps the function composition around to match.